CopyDocumentModal component added
